* Fix - moved Factory classes to dedicated package. * Fix - moved Suite related models classes to new suite package in model package.
