+++ testApi/src/main/scala/africa/absa/testing/scapi/model/suite/SuiteBefore.scala
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package africa.absa.testing.scapi.model
+package africa.absa.testing.scapi.model.suite
 
 /**
- * Case class that represents a suite after methods.
+ * Case class that represents a suite before methods.
  *
- * @param name The name of the after methods.
- * @param methods The set of suite after methods.
+ * @param name The name of the before methods.
+ * @param methods The set of suite before methods.
  */
-case class SuiteAfter(name: String, methods: Set[Method]) extends SuiteAround(name, methods) {
+case class SuiteBefore(name: String, methods: Set[Method]) extends SuiteAround(name, methods) {
   override def resolveReferences(references: Map[String, String]): SuiteAround = {
-    SuiteAfter(
+    SuiteBefore(
       name,
       methods.map(c => c.resolveReferences(references))
     )
